+++ .gitignore
@@ -1,1 +1,1 @@
-build/
+/build
