+++ go.mod
@@ -2,5 +2,4 @@ module github.com/flared/lokify
 
 go 1.16
 
-require github.com/gorilla/mux v1.8.0 // indirect
-
+require github.com/gorilla/mux v1.8.0
